[lib] Make api-request-handler keyserver-specific
Summary:
Select a connection status from the given keyserver.
Modify `socketAPIHandlers` to be a map from keyserver to `SocketAPIHandler`.

Test Plan: Added console.logs in `fetchResponse` to check if it is called and if a request is matched with response. Focused a thread and checked if proper request and response were generated.

Reviewers: inka, kamil, michal, atul

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11089