[lib] Make activity-handler keyserver-specific
Summary:
Add a prop that identifies a keyserver.
Make sure that we queue only the updates to a keyserver that owns a thread.

Depends on D11089

Test Plan:
Focus a thread and check if the action is dispatched.
Using React debugger, modified `ActionHandler`'s `activeThread` prop to a thread that has a different prefix - that resulted in only one, `focus: false` update being generated. Changed it back to the correct prefix - again, only one update was generated (this time `focus: true`).

Reviewers: inka, kamil, michal, atul