[native] Reintroduce `createUpdateDBOpsForThreadStoreThreadInfos` for new `RawThreadInfo` encoding
Summary:
Discussed in D11057, but we want to "freeze" the previous implementation of `createUpdateDBOpsForThreadStoreThreadInfos` in place to support old migrations while re-introducing a new version that only handles "MinimallyEncoded -> MinimallyEncoded` migrations. We only have "MinimallyEncoded -> MinimallyEncoded` migrations in the future, so IMO it doesn't make sense to clutter up `createUpdateDBOpsForThreadStoreThreadInfos` to be able to handle

- LegacyRawThreadInfos => LegacyRawThreadInfos
- LegacyRawThreadInfos => MinimallyEncodedRawThreadInfos
- MinimallyEncodedRawThreadInfos => MinimallyEncodedRawThreadInfos

It also lets us avoid having to go back and test some of the old migrations when we tweak `createUpdateDBOpsForThreadStoreThreadInfos`.

---

Depends on D11057

Test Plan: This gets consumed for the first time in the upcoming specialRoles migration where it will be tested implicitly.

Reviewers: ginsu, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11095