[web] Only open tunnelbroker on the active tab
Summary:
[ENG-6668 : Only open tunnelbroker on a tab with a lock](https://linear.app/comm/issue/ENG-6668)

Using the previously `useWebLock` hook, we can open the tunnelbroker connection only on one of the tabs in the foreground.

Depends on D11114

Test Plan:
Switch between different tabs, make sure after a little bit of time (disconnetion of the previous tab and connection of the current tab) the new tabs connects to the tunnelbroker. Make sure that if multiple tabs are visible and in the foreground, only one tab connects to the tunnelbroker.

Reviewers: kamil, inka, bartek

Subscribers: