[web] implementing publishing prekeys to Identity
Summary:
This code is web specific - so I think it's butter to implement it this way than adding this to `IdentityServiceClient` and mocking on native.
On native this method is used only from C++ level and is not exposed to JS.

Depends on D11109

Test Plan:
Call this method and check in staging DDB that prekeys were updated.
```
await publishPrekeysToIdentity({
  contentPrekey: 'test0',
  contentPrekeySignature: 'test1',
  notifPrekey: 'test2',
  notifPrekeySignature: 'test3',
}, identityClient);
```

Reviewers: marcin, varun, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11110