[lib] Modify update calendar query so that it can update a subset of keyservers
Summary: `CalendarQueryHandler` will be dispatching this action for a single keyserver - we need to be able to do that. Later in the stack I'm going to move `actualizedCalendarQuery` into `keyserverStore` so that it becomes keyserver-specific.

https://linear.app/comm/issue/ENG-5779/update-calendarqueryhandler

Depends on D11094

Test Plan: Modify date range on a calendar and check if the query is correctly updated.
Modify the code by setting an explicit keyserver id and check if the solution still works and a keyserver ID is present in the reducer.

Reviewers: inka, kamil, michal, atul