[backup service] Revoke log blob holders
Summary:
My mistake - forgot to add this code to the backup service while adding logs. This diff adds the code to call blob service revoke holder after removing a log. It also improves the holder revoke for backups a bit so that the blob is called inside of the `remove_backup_item` function and not in the http handler.

Depends on D11142

Test Plan:
- Create a new backup with logs
- Make sure that the log is stored in blob service (and not inline in dynamodb)
- Check that the log blob has a holder
- Delete the backup
- Check the the log blob no longer has a holder

Reviewers: bartek, kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11143