[lib] Move actualizedCalendarQuery into keyserver store
Summary:
Move the filed to the store - this change means that we keep a separate value for each keyserver. This value includes only the filters from that keyserver.

https://linear.app/comm/issue/ENG-6901/move-actualizedcalendarquery-into-keyserver-store

Depends on D11135

Test Plan:
Check if a migration succeeds on native and web - inspect Redux store after the migration and check if there are no inconsistencies reported.
Open a calendar, modify date range and check if the value is correctly updated in keyserver store.
Close and reopen the native app - there should be no inconsistency warning.
Log out and in, modify the dates - the store should be updated correctly.

Reviewers: michal, atul, inka, kamil

Reviewed By: inka, kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11144