[terraform] ddb table for identity workflows in progress
Summary: we've had some issues with sticky sessions on web. specifically, grpc-web seems to ignore the load balancer cookie, which we rely on for stickiness. we need sticky sessions because some workflows require multiple RPCs, and only the identity service node that received the first RPC will have the necessary info to handle the subsequent ones. if we write this "state" to a DynamoDB table, any node can access it, eliminating the need for stickiness altogether.

Test Plan: terraform apply works on staging

Reviewers: will

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11150