[lib][keyserver] Delete MORE_ONE_TIME_KEYS mechanism
Summary: We no longer rely on keyserver to keep the keys - this is a responsibility of Identity service.

https://linear.app/comm/issue/ENG-6521/check-serverrequesttypesmore-one-time-keys

Depends on D11176

Test Plan:
Check if keyserver DB migration deletes the table.
Check if it is possible to run web and native client.
Run an older native client and check if it works with new keyserver.

Reviewers: michal, atul, inka, kamil