[lib] Use keyserverID when generating initial message using socket
Summary:
Use the actual `keyserverID`.

https://linear.app/comm/issue/ENG-6523/use-keyserver-id-when-generating-a-response-to

Depends on D11177

Test Plan:
Delete a session on keyserver, restart the native client and verify that the code handling `INITIAL_NOTIFICATIONS_ENCRYPTED_MESSAGE` was called. Send a message from web client and verify that a notification was delivered without a warning about missing session.

Reviewers: michal, atul, marcin, kamil