[lib] Compute hashes only from one keyserver
Summary:
When doing state check we should only compute hashes for entities from a keyserver performing the check.

https://linear.app/comm/issue/ENG-3915/refactor-generating-a-response-to-serverrequesttypescheck-state

Depends on D11183

Test Plan:
Checked that the state sync still works - modified a thread and an entry and verified that the sync updates values properly on the client side.
Modified state selectors in entries and threads specs to return one more entity from a different keyserver and verified that the hashes were still correct (also performed the test mentioned earlier with the modified code).

Reviewers: michal, atul, inka, kamil