[lib] Replaces the getAuthmetadata conditional with loggedIn conditional
Summary:
This replaces the getAuthMetadata conditional (which wasn't necessary) with whether or not we're logged in. If we're not logged in,
we should return early with a null as to avoid checking the contents of the auth metadata

Downstream this will result in the IdentitySearchAccessToken always being set to null, causing the identity search context to not initialize.
As users are not logged in, there shouldn't be any need to use identity search anyway.

Test Plan: Tested on local. Successfully returned a null IdentitySearchAccessToken value

Reviewers: varun

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11200