[web] Compute `existingRole` instead of `existingRoleID` in `RoleActionsMenu`
Summary:
This should be a noop refactor that I'm separating into separate diff to make next diff easier to remove.

We basically grab the `RoleInfo` instead of the `RoleInfo.id` so that we can access the `specialRole` field and deprecate `useRoleDeletableAndEditableStatus`.

---

Depends on D11205

Test Plan:
`flow`, the `invariant`, logging the value, making sure `RoleActionsMenu` continues to look/work as expected

Reviewers:
ginsu, tomek, kamil, michal