[lib][native] Introduce supportingMultipleKeyservers constant
Summary:
Introducing this constant right now even though it's not really doing anything.

As mentioned a couple diffs back, I want this constant introduced so we can be clear on the various distinct launches we'll have in 2024.

The code in SQLiteDataHandler doesn't yet support recovery from a corrupt SQLite database in a multi-keyserver scenario, as this depends on restoring from the backup service.

Question for @michal: is there an existing task tracking implementing backup restore for corrupt database recovery, for both native and web? I want to add a comment referencing this code in `SQLiteDataHandler`, to make sure it's updated as part of that task.

Depends on D11229

Test Plan: I used this test plan for the whole stack: https://gist.github.com/Ashoat/75ab690d5c53cdd68a51b02e03e27c58

Reviewers: michal, inka, tomek

Subscribers: