[keyserver] Transform client IDs when handling socket message
Summary: When receiving a socket message, the server wasn't transforming the IDs - it caused `verifyCalendarQueryThreadIDs` to throw because none of the received thread filter IDs were present in the DB.

https://linear.app/comm/issue/ENG-7115/changing-calendar-query-doesnt-work

Test Plan: Add and remove some thread filters and check if the error message no longer occurs. Also check if the `actualizedCalendarQuery` is correctly updated in the store.

Reviewers: atul, inka, michal, marcin