[lib] Remove `isDefault` field from `RoleInfo`
Summary:
We previously removed all usages of `RoleInfo.isDefault` on the `native`/`web` client, so we can safely remove it from `RoleInfo`.

As usual, there will be a `native`/`web` migration to go along with this change. Will actually update the existing `specialRole` migration in place since it makes sense to couple the inclusion of `specialRole` and exclusion of `isDefault` in a single pass.

---

Depends on D11252

Test Plan:
`flow` + verified that `RoleInfo`s sent from the `keyserver` have the `isDefault` field excluded.

Did a search through the codebase to ensure there were not any instances of `isDefault` that are relevant to changing this client type.

Reviewers: ginsu, tomek, michal, kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11253