[identity] db client methods to interact with workflows table
Summary:
methods to insert and get workflows from the workflows table. the insert method returns a UUID that will be returned to the gRPC client caller. we don't need a remove method because the TTL is enabled.

Depends on D11160

Test Plan: called the insert and get methods and got back the same WorkflowInProgress

Reviewers: bartek, will, michal

Subscribers: