[keyserver][native] remove cookie logic
Summary:
the identity service is now completely stateless again so we can safely remove cookie logic from our gRPC clients. this code ensured that the same identity service node would handle start and finish RPCs, but that is no longer necessary since we're storing the in-progress workflows in DDB instead of in memory. this code was only added to the gRPC clients, not gRPC-web, so there is nothing to remove from web.

Depends on D11256

Test Plan: successfully logged in and registered from native and keyserver with staging identity

Reviewers: bartek, will, michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11257