[keyserver] Print warnings if usingIdentityCredentials not set
Summary:
issue: [[https://linear.app/comm/issue/ENG-6618/make-yarn-dev-print-an-error-informing-the-user-that-they-need-to | ENG-6618]]
We want to warn the dev that they are not using identity credentials for communictaing with the clients

For now using identity credentials is optional. The dev can proceed with their setup. If they want to start using the identity credentials though, they need to restart nix and go through setting up a new keyserver.
If the keyserver is run in Docker, the dev should set `usingIdentityCredentials` flag in `COMM_JSONCONFIG_secrets_user_credentials` env variable and create a new keyserver. See [[https://www.notion.so/commapp/Running-two-keyservers-4295f98e7b0547d4ba027ba52c2d2e80 | doc]]

We don't need this warning on the prod keyserver - it will not have the flag set, but that's fine, because the dafualt `256` is its id.

Test Plan: Ran the keyserver and checked that if `usingIdentityCredentials` is set then the warning is not printed, and if it is set, then the warning is printed

Reviewers: tomek, michal

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11267