[keyserver] Instruct to restart nix if login / register fails in the new flow
Summary:
issue: [[https://linear.app/comm/issue/ENG-7137/print-a-warning-informing-the-user-that-they-need-to-restart-nix-if | ENG-7137]]
If the dev is using the new multikeyserver flow, they had provided some credentials during when nix was run. If registration with these credentials failed, they neeed to rerun nix, so that they can provide new credentials, but also so that the db gets fixed.

Test Plan:
Tested that if registration fails, and the flag is set, the warning is printed

Reviewers: tomek, michal

Subscribers: