[lib] Don't sync users with non-authoritative keyservers
Summary: We can take a similar approach to what we do with threads that are being hashed - inform the server that there are no inconsistencies. This is a kind of hack, but it will be later improved when we start syncing the store between clients - then the keyserver will no longer attempt the sync.

https://linear.app/comm/issue/ENG-7132/stop-performing-state-sync-of-users-and-current-user-with-non

Test Plan: Check if the state sync still works for one keyserver - modify keyserver DB and check if the state is fixed.
Connect to the second keyserver and check if the state difference doesn't cause an update on client side.

Reviewers: inka, kamil, michal, varun