[identity] add optional force param to password login RPC
Summary:
[context here](https://linear.app/comm/issue/ENG-6921/limit-users-to-1-keyserver)

tl;dr we neeed an optional force param so that we can determine when a user is intentionally replacing their existing keyserver with a new one

Test Plan: logged in to local identity from local keyserver. then tried logging in from docker keyserver with the same credentials -- this failed. then tried logging in again from the docker keyserver with the same credentials AND `force` set to true -- this succeeded. The old keyserver device id was removed from DDB and the new one was added

Reviewers: bartek

Reviewed By: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11278