[keyserver] Create json files with authoritative keyserver id
Summary:
issue: [[ https://linear.app/comm/issue/ENG-7136/create-authoritative-keyserverjson-files-when-the-keyserver-is-run | ENG-7136 ]]
These files are used to configure the authoritative keyservers data on the clients. They are created by the keyserver, because to know the id the keyserver has to first log in.

One effect of this, is that the keyserver will have to be run before web and native the first time. I should mention that in the [[https://www.notion.so/commapp/Running-two-keyservers-4295f98e7b0547d4ba027ba52c2d2e80 | doc]].  I made a note to do that in [[https://linear.app/comm/issue/ENG-7173/update-the-testing-doc | ENG-7173 ]]

Test Plan: Tested that the files are correctly created when the keyserver is run

Reviewers: tomek, michal

Reviewed By: michal

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11282