[terraform] use identity 0.12 on prod
Summary: this is the same image as 0.12-staging, just re-tagged for prod. after this is approved i will run terraform apply.

Test Plan:
tested this image extensively on staging last week.

after applying this change we can remove the access token and user ID from the production mariadb metadata table and restart the keyserver. the table should be repopulated with the same user ID and a new access token.

Reviewers: ashoat, bartek, will

Reviewed By: ashoat, bartek, will

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D11305