[native][web] Update `ChatInputBar` to pass `communityThreadInfo` to `threadFrozenDueToViewerBlock`
Summary:
We added `communityThreadInfo: ?(RawThreadInfo | ThreadInfo)` field to `threadFrozenDueToViewerBlock` and updated `ChatInputBar` on both clients to pass in `communityThreadInfo`. For now we're not doing anything with `communityThreadInfo` in `threadFrozenDueToViewerBlock`, this is in preparation for that. We need to do something similar for `threadFrozenDueToBlock` and then `threadIsWithBlockedUserOnly` and then `threadOrParentThreadHasAdminRole`, etc.

---

Depends on D11286

Test Plan: Logged `communityThreadInfo` in `ChatInputBar` on clients + in `threadFrozenDueToViewerBlock` and ensured that it appeared as expected.

Reviewers: tomek, ginsu, kamil, michal

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11308