[web] olmAPI on shared worker
Summary:
[ENG-6768 : Migrate validateAndUploadPreKeys to shared worker](https://linear.app/comm/issue/ENG-6768)
[ENG-6658 : Migrate encrypt, decrypt to shared worker](https://linear.app/comm/issue/ENG-6658)
Part of [ENG-6657 : Migrate session creation code to shared worker](https://linear.app/comm/issue/ENG-6657)

Using a similar approach as in D11275 (one message type for calling the methods) this diff moves the `olmAPI` to the shared worker. There are two logic changes which I have annotated in inline comments. Other than that the only changes are related to using real persisting functions and not the mocked ones in `olm-api.js` (which are now removed). In this case the `olmAPI` proxied functions are not hidden behind a flag as they were using mock functions anyway.

Depends on D11326

Test Plan:
Called each of the methods and checked if they work. The `encrypt`/`decrypt`/`contentInboundSessionCreator` have been tested in context of peer to peer message communication.

Reviewers: kamil, tomek

Subscribers: